refactored files and project structure
