+++ src/components/expense/Expenses.tsx
@@ -5,8 +5,8 @@ import { useEffect, useState } from "react";
 import { ExpenseService } from "@/service/ExpenseService";
 import closeSvg from '../../assets/close.svg'
 import Image from "next/image";
-import RowActions from "../RowActions";
-import FilterSelector from "../FilterSelector";
+import RowActions from "../shared/RowActions";
+import FilterSelector from "../shared/FilterSelector";
 import { filterItems } from "@/util/utils";
 import { useSelector } from "react-redux";
 import { budgetSelectors } from "@/store";
