+++ src/components/groceries/Groceries.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from "react";
 import GroceryItemForm from "./GroceryItemForm";
 import { GroceryService } from "@/service/GroceryService";
 import Image from "next/image";
-import RowActions from "../RowActions";
-import FilterSelector from "../FilterSelector";
+import RowActions from "../shared/RowActions";
+import FilterSelector from "../shared/FilterSelector";
 import { filterItems } from "@/util/utils";
 import { useSelector } from "react-redux";
 import { budgetSelectors } from "@/store";
