+++ src/components/income/Income.tsx
@@ -3,8 +3,8 @@ import {  IncomeItem } from "@/model/models";
 import IncomeItemForm from "./IncomeItemForm";
 import { useEffect, useState } from "react";
 import { IncomeService } from "@/service/IncomeService";
-import RowActions from "../RowActions";
-import FilterSelector from "../FilterSelector";
+import RowActions from "../shared/RowActions";
+import FilterSelector from "../shared/FilterSelector";
 import { filterItems } from "@/util/utils";
 import Image from "next/image";
 import closeSvg from '../../assets/close.svg';
