+++ src/components/investment/Investments.tsx
@@ -3,8 +3,8 @@ import {  InvestmentItem } from "@/model/models";
 import InvestmentItemForm from "./InvestmentItemForm";
 import { useEffect, useState } from "react";
 import { InvestmentService } from "@/service/InvestmentService";
-import RowActions from "../RowActions";
-import FilterSelector from "../FilterSelector";
+import RowActions from "../shared/RowActions";
+import FilterSelector from "../shared/FilterSelector";
 import { filterItems } from "@/util/utils";
 import Image from "next/image";
 import closeSvg from '../../assets/close.svg';
