+++ src/components/investment/dashboard-charts/InvestmentsLineBarPanel.tsx
